+++ packages/dflex-dnd/src/LayoutManager/constants.ts
